+++ services/dns/mgmt/2016-04-01/dns/CHANGELOG.md
@@ -1,2 +1,18 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. ZoneType.Private
+1. ZoneType.Public
+
+### New Funcs
+
+1. PossibleZoneTypeValues() []ZoneType
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. ZoneProperties.ZoneType